requirements.txt
langchain>=0.3.27
langchain_openai>=0.3.35
langgraph>=0.6.10
openai>=2.3.0
python-dotenv>=1.1.1
requests>=2.32.5
streamlit>=1.49.1
typing_extensions>=4.15.0
